Create category page was created, and view users
View users page was updated to include FAwesome icon avatar and better layout with personal info.
Index page was updated to show recent categories (newest first) if user is logged in.

Forms were updated to have a more global styling applied to field areas.

User bar was updated to show if user was logged in or not.